Support GitHub Updater
The GitHub Updater plugin allows plugins to be updated directly from GitHub, just by including a couple of commented lines within the plugin header.

The standard is for GitHub Updater to check for tags, so you should tag each release. If there are no tags, then it will check the branch mentioned (here, master). As such, you should create a branch from master called develop and do all future PRs to that (make developer the default branch via this repo's Settings in the sidebar), which leaves master as the clean / stable version. You're now following the basics of git-flow!